hass: increase WebSocket buffer size
When using a Home Assistant intent like "how many lights are on", the
response contains all light entities. With the default buffer size of
1024, this breaks the JSON in chunks that are no longer parseable.

Increase the buffer size to 8192. This roughly double of what is needed
for the aformentioned intent on my Home Assistant setup. This value is
larger than CONFIG_SPIRAM_MALLOC_ALWAYSINTERNAL=4096 so should always be
allocated in SPIRAM, so we could even increase this value further.

Short torture test of 80 runs with a few more tweaks in sdkconfig shows
extremely stable internal memory usage:

    free 134007 allocated 168592 min_free 133659 largest_free_block 122880

If we still have a leak, it will happen with SPIRAM memory and it will
take a long time before hitting it, but this should at least help with
the play_tone task failing with the following error after some uptime:

AUDIO_THREAD: Error creating RestrictedPinnedToCore play_tone_ok